pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orby-cli"
version = "0.1.0"
description = "The local-first AI CLI with agentic powers. Connect Ollama, LM Studio, Hugging Face — no cloud required."
readme = "README.md"
authors = [{name = "Orby Team"}]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
]
dependencies = [
    "click>=8.0.0",
    "requests>=2.28.0",
    "aiohttp>=3.8.0",
    "prompt-toolkit>=3.0.0",
    "pyyaml>=6.0",
    "colorama>=0.4.0",
    "inquirer>=3.0.0",
    "transformers>=4.30.0",
    "torch>=2.0.0; sys_platform != 'darwin'",
    "torch>=2.0.0; sys_platform == 'darwin'",
    "rich>=13.0.0",
    "textual>=0.40.0",
    "sentence-transformers>=2.2.0",
    "PyPDF2>=3.0.0",
    "python-docx>=0.8.0",
    "beautifulsoup4>=4.12.0",
    "watchdog>=3.0.0"
]

[project.optional-dependencies]
voice = [
    "whisper>=1.0.0"
]
vision = [
    "pytesseract>=0.3.10",
    "pillow>=9.0.0"
]
web = [
    "streamlit>=1.20.0"
]

[project.scripts]
orby = "orby.cli:main"

[project.urls]
Homepage = "https://github.com/orby/orby"
Repository = "https://github.com/orby/orby"